+++ shopcrafter-service/src/main/java/ru/ilya/shopcraftercore/service/goods/ProductService.java
@@ -2,9 +2,12 @@ package ru.ilya.shopcraftercore.service.goods;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import ru.ilya.shopcraftercore.dto.goods.product.ProductDto;
 import ru.ilya.shopcraftercore.dto.goods.product.PutProductDto;
+import ru.ilya.shopcraftercore.entity.goods.Category;
 import ru.ilya.shopcraftercore.entity.goods.Product;
+import ru.ilya.shopcraftercore.repository.goods.CategoryRepository;
 import ru.ilya.shopcraftercore.repository.goods.ProductRepository;
 
 import java.util.List;
@@ -12,36 +15,62 @@ import java.util.List;
 @Service
 public class ProductService {
     private final ProductRepository productRepository;
+    private final CategoryRepository categoryRepository;
 
     @Autowired
-    public ProductService(ProductRepository productRepository) {
+    public ProductService(ProductRepository productRepository, CategoryRepository categoryRepository) {
         this.productRepository = productRepository;
+        this.categoryRepository = categoryRepository;
     }
 
-    public List<ProductDto> getAllProducts() {
-        return ((List<Product>) productRepository.findAll())
-                .stream().map(ProductDto::fromEntity)
-                .toList();
+    public List<ProductDto> getAllProducts(long storeId, long categoryId) {
+        Category category = categoryRepository.findByStoreIdAndId(storeId, categoryId);
+        if (category == null) {
+            throw new RuntimeException("Category not found");
+        }
+        List<Product> products = productRepository.findByCategoryId(categoryId);
+        if (products == null) {
+            return null;
+        }
+        return products.stream().map(ProductDto::fromEntity).toList();
     }
 
-    public ProductDto getProductById(long id) {
-        Product product = productRepository.findById(id).orElse(null);
+    public ProductDto getProductById(long storeId, long categoryId, long id) {
+        Category category = categoryRepository.findByStoreIdAndId(storeId, categoryId);
+        if (category == null) {
+            throw new RuntimeException("Category not found");
+        }
+
+        Product product = productRepository.findByCategoryIdAndId(categoryId, id);
         if (product == null) {
             return null;
         }
         return ProductDto.fromEntity(product);
     }
 
-    public ProductDto createProduct(PutProductDto putProductDto) {
+    @Transactional
+    public ProductDto createProduct(long storeId, long categoryId, PutProductDto putProductDto) {
+        Category category = categoryRepository.findByStoreIdAndId(storeId, categoryId);
+        if (category == null) {
+            throw new RuntimeException("Category not found");
+        }
         Product product = new Product();
         product.setName(putProductDto.getName());
+        product.setCategory(category);
+        category.getProducts().add(product);
         product.setDescription(putProductDto.getDescription());
         productRepository.save(product);
+        categoryRepository.save(category);
         return ProductDto.fromEntity(product);
     }
 
-    public ProductDto updateProduct(long id, PutProductDto putProductDto) {
-        Product product = productRepository.findById(id).orElse(null);
+    @Transactional
+    public ProductDto updateProduct(long storeId, long categoryId, long id, PutProductDto putProductDto) {
+        Category category = categoryRepository.findByStoreIdAndId(storeId, categoryId);
+        if (category == null) {
+            throw new RuntimeException("Category not found");
+        }
+        Product product = productRepository.findByCategoryIdAndId(categoryId, id);
         if (product == null) {
             return null;
         }
@@ -51,7 +80,18 @@ public class ProductService {
         return ProductDto.fromEntity(product);
     }
 
-    public void deleteProduct(long id) {
+    @Transactional
+    public void deleteProduct(long StoreId, long CategoryId, long id) {
+        Category category = categoryRepository.findByStoreIdAndId(StoreId, CategoryId);
+        if (category == null) {
+            throw new RuntimeException("Category not found");
+        }
+        Product product = productRepository.findByCategoryIdAndId(category.getId(), id);
+        if (product == null) {
+            throw new RuntimeException("Product not found");
+        }
+        category.getProducts().remove(product);
+        categoryRepository.save(category);
         productRepository.deleteById(id);
     }
